requirements.txt
streamlit
snowflake
snowflake-snowpark-python
geopandas
plotly
python-dotenv